requirements.txt
Flask>=2.2
SQLAlchemy>=1.4
Flask-SQLAlchemy>=3.0
Flask-JWT-Extended>=4.4
argon2-cffi>=21.1.0
marshmallow>=3.14
Flask-Limiter>=2.8
Flask-Talisman>=1.0
python-dotenv>=1.0
Flask-Migrate>=4.0
pytest>=7.0
pytest-cov>=4.0
itsdangerous>=2.1
